+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # 1.0.0 (2025-12-15)
 
-
 ### Features
 
-* initial release - unstorage cache ([ec36969](https://github.com/dschewchenko/drizzle-uncache/commit/ec36969b9eea432052cdfdfb3bb7942a56681e0f))
+- initial release - unstorage cache ([ec36969](https://github.com/dschewchenko/drizzle-uncache/commit/ec36969b9eea432052cdfdfb3bb7942a56681e0f))
 
 # Changelog